add gaussian base class
