+++ floris/simulation/wake_velocity/gaussianModels/gauss_legacy.py
@@ -16,7 +16,7 @@ from ..base_velocity_deficit import VelocityDeficit
 from .gaussian_model_base import GaussianModel
 
 
-class LegacyGauss(VelocityDeficit):
+class LegacyGauss(GaussianModel):
     default_parameters = {
         'ka': 0.38,
         'kb': 0.004,
@@ -29,16 +29,12 @@ class LegacyGauss(VelocityDeficit):
     }
 
     def __init__(self, parameter_dictionary):
-        print(parameter_dictionary)
+
         super().__init__(parameter_dictionary)
         self.logger = setup_logger(name=__name__)
 
-        print(self.calculate_VW_velocities)
-
         self.model_string = "gauss_legacy"
         model_dictionary = self._get_model_dict(__class__.default_parameters)
-
-        print(self.calculate_VW_velocities)
 
         # near wake / far wake boundary parameters
         self.alpha = model_dictionary["alpha"]
@@ -47,6 +43,13 @@ class LegacyGauss(VelocityDeficit):
         # wake expansion parameters
         self.ka = model_dictionary["ka"]
         self.kb = model_dictionary["kb"]
+
+        # GCH Parameters
+        self.calculate_VW_velocities = model_dictionary["calculate_VW_velocities"]
+        self.use_yaw_added_recovery = model_dictionary["use_yaw_added_recovery"]
+        self.yaw_recovery_alpha = model_dictionary["yaw_recovery_alpha"]
+        self.eps_gain = model_dictionary["eps_gain"]
+
 
     def function(self, x_locations, y_locations, z_locations, turbine, turbine_coord, deflection_field, flow_field):
         # veer (degrees)
